+++ CHANGELOG.md
@@ -21,6 +21,10 @@
 * Added elaboration for accesses to fields of struct constants, which can
   substantially improve performance on some designs
 * Added constant folding for comparisons involving string literals
+* Port connection attributes (e.g., [pulp_soc.sv]) are now ignored with a
+  warning rather than failing to parse
+
+[pulp_soc.sv]: https://github.com/pulp-platform/pulp_soc/blob/0573a85c/rtl/pulp_soc/pulp_soc.sv#L733
 
 ## v0.0.10